Rename sources to view
