+++ projects/plugins/jetpack/extensions/blocks/map/view/index.js
